milestone 0,1
Design the structure of the global state along the lines of the exercises carried out over the days
precedents.

Create a basic layout with a searchbar (an input and a button) where we can
write the name of a film completely or partially. We can, by clicking the
button, search the API for all the films that contain what the user wrote.
After the API response, we want to display the following values for each on the screen
movie found:
1. Title
2. Original Title
3. Language
4. Voting