milestones 2
We transform the static language string into a real language flag
corresponding nation, handling the case where we do not have the flag of
country returned by the API (the flags are not there in FontAwesome).
We then broaden the search to include TV series. With the same search action
we will have to take both the films that match the query and the TV series, being
be careful to have similar values at the end (series and films have fields in the JSON of
different answer, similar but not always identical)